+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 - Visualization for categories of open habits
 - Category name validation checking for duplicates
+- Categories filter bottom sheet with categories
 
 ### Changed:
 - Upgraded dependencies